backend/requirements.txt
aioftp==0.20.1
aiohttp==3.9.3
aiohttp-apispec==3.0.0b2
aiohttp-jinja2==1.5.1
aiohttp-security==0.4.0
aiohttp-session==2.12.0
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
apispec==6.6.1
appnope==0.1.4
asttokens==2.4.1
asyncssh==2.14.1
attrs==23.2.0
Authlib==1.3.1
azure-ai-translation-document==1.0.0
azure-core==1.31.0
Babel==2.15.0
bandit==1.7.9
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
coverage==7.5.3
crefi==2.0.9
cryptography==42.0.2
cssselect2==0.7.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
dirhash==0.2.1
distlib==0.3.8
dnspython==2.1.0
docutils==0.18.1
dparse==0.6.4b0
elastic-transport==8.15.0
elasticsearch==8.15.0
email_validator==2.1.1
executing==2.1.0
Faker==28.0.0
fastapi==0.111.0
fastapi-cli==0.0.2
filelock==3.15.1
Flask==3.0.3
fonttools==4.51.0
fpdf2==2.7.8
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
identify==2.5.36
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.7.2
ldap3==2.9.1
lxml==4.9.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
msrest==0.7.1
multidict==6.0.5
myst-parser==2.0.0
nest-asyncio==1.6.0
nodeenv==1.9.1
oauth2client==4.1.3
oauthlib==3.2.2
orjson==3.10.3
packaging==23.2
parso==0.8.4
pathspec==0.12.1
pbr==6.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pyproject-api==1.6.1
pytest==8.2.2
pytest-aiohttp==1.0.3
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyxattr==0.8.1
PyYAML==6.0.1
pyzmq==26.2.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.2.3
safety-schemas==0.0.2
scantree==0.0.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.1.2
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
stack-data==0.6.3
starlette==0.37.2
stevedore==5.2.0
svglib==1.5.1
tinycss2==1.3.0
tornado==6.4.1
tox==4.15.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
utils==1.0.2
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.21.0
wcwidth==0.2.13
webargs==8.4.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.4
wheel==0.43.0
yarl==1.9.4
